feat: send email when team is registered

+++ app/Http/Controllers/TeamController.php
@@ -8,12 +8,18 @@ use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Mail;
 use Inertia\Inertia;
 
 class TeamController extends Controller
 {
   public function create(Request $request)
   {
+    if (
+      \Carbon\Carbon::parse(env('END_DATE'))
+      ->lt(\Carbon\Carbon::now('Asia/Kolkata'))
+    ) return redirect('/dashboard');
+
     $body = $request->validate([
       'email' => 'required|email',
       'student_1' => 'required',
@@ -29,11 +35,15 @@ class TeamController extends Controller
       return redirect('/');
     }
 
+    $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+    $password = '';
+    for ($i = 0; $i < 15; $i++) {
+      $password = $password . $characters[rand(0, strlen($characters))];
+    }
     $user = new User([
       'type' => 'team',
       'email' => $body['email'],
-      // TODO randomly generate and email
-      'password' => Hash::make('password'),
+      'password' => Hash::make($password),
     ]);
     try {
       $user->save();
@@ -50,11 +60,20 @@ class TeamController extends Controller
     ]);
     $teamInfo->save();
 
+
+    Mail::to($user)
+      ->send(new \App\Mail\TeamCreatedMail($user, $password));
+
     return redirect("/");
   }
 
   public function destroy(User $team)
   {
+    if (
+      \Carbon\Carbon::parse(env('START_DATE'))
+      ->lt(\Carbon\Carbon::now('Asia/Kolkata'))
+    ) return redirect('/dashboard');
+
     $team->delete();
 
     return redirect('/');
